requirements.txt
# main
nonebot2~=2.0.0
nonebot-adapter-onebot~=2.2.3

# # tests
# pytest~=7.1.2

# greeting
requests>=2.31.0
requests-html~=0.10.0

# repeater
pydantic~=1.10.0
pymongo~=4.3.3
jieba~=0.42.1
pypinyin~=0.49.0
